+++ src/command_modules/azure-cli-resource/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+2.1.1
+++++++
+* `group deployment create`: Add `--rollback-on-error` to execute a known-good deployment on error.
+
 2.1.0
 +++++
 * BREAKING CHANGE: 'show' commands log error message and fail with exit code of 3 upon a missing resource.